chore: expose JSON Reporter types
