+++ utils/generate_types/index.js
@@ -566,7 +566,15 @@ class TypesGenerator {
     const generator = new TypesGenerator({
       documentation,
       classNamesToGenerate: new Set(reporterDocumentation.classesArray.map(cls => cls.name)),
-      ignoreMissing: new Set(['FullResult']),
+      ignoreMissing: new Set([
+        'FullResult',
+        'JSONReport',
+        'JSONReportSuite',
+        'JSONReportSpec',
+        'JSONReportTest',
+        'JSONReportTestResult',
+        'JSONReportTestStep',
+      ]),
       includeExperimental,
     });
     return await generator.generateTypes(path.join(__dirname, 'overrides-testReporter.d.ts'));
